Move SQL Alchemy models out of storage adapter for #904
