+++ chatterbot/storage/sql_storage.py
@@ -1,108 +1,10 @@
 import random
-
 from chatterbot.storage import StorageAdapter
-from chatterbot.conversation import Response
-from chatterbot.conversation import Statement
 
 try:
-    from chatterbot.ext.sqlalchemy_app.models import Base
-    from sqlalchemy.orm import relationship
-    from sqlalchemy.sql import func
-    from sqlalchemy import Table, Column, Integer, String, DateTime, ForeignKey, PickleType
-
-    tag_association_table = Table(
-        'tag_association',
-        Base.metadata,
-        Column('tag_id', Integer, ForeignKey('tag.id')),
-        Column('statement_id', Integer, ForeignKey('StatementTable.id'))
+    from chatterbot.ext.sqlalchemy_app.models import (
+        Base, Conversation, StatementTable, ResponseTable, Tag
     )
-
-    class Tag(Base):
-        """
-        A tag that describes a statement.
-        """
-
-        name = Column(String)
-
-    class StatementTable(Base):
-        """
-        StatementTable, placeholder for a sentence or phrase.
-        """
-
-        __tablename__ = 'StatementTable'
-
-        def get_statement(self):
-            statement = Statement(self.text, extra_data=self.extra_data)
-            for response in self.in_response_to:
-                statement.add_response(response.get_response())
-            return statement
-
-        text = Column(String, unique=True)
-
-        tags = relationship(
-            'Tag',
-            secondary=lambda: tag_association_table,
-            backref='statements'
-        )
-
-        extra_data = Column(PickleType)
-
-        in_response_to = relationship(
-            'ResponseTable',
-            back_populates='statement_table'
-        )
-
-    class ResponseTable(Base):
-        """
-        ResponseTable, contains responses related to a givem statment.
-        """
-
-        __tablename__ = 'ResponseTable'
-
-        text = Column(String)
-
-        created_at = Column(
-            DateTime(timezone=True),
-            server_default=func.now()
-        )
-
-        occurrence = Column(Integer, default=1)
-
-        statement_text = Column(String, ForeignKey('StatementTable.text'))
-
-        statement_table = relationship(
-            'StatementTable',
-            back_populates='in_response_to',
-            cascade='all',
-            uselist=False
-        )
-
-        def get_response(self):
-            occ = {'occurrence': self.occurrence}
-            return Response(text=self.text, **occ)
-
-    conversation_association_table = Table(
-        'conversation_association',
-        Base.metadata,
-        Column('conversation_id', Integer, ForeignKey('conversation.id')),
-        Column('statement_id', Integer, ForeignKey('StatementTable.id'))
-    )
-
-    class Conversation(Base):
-        """
-        A conversation.
-        """
-
-        __tablename__ = 'conversation'
-
-        id = Column(Integer, primary_key=True)
-
-        statements = relationship(
-            'StatementTable',
-            secondary=lambda: conversation_association_table,
-            backref='conversations'
-        )
-
 except ImportError:
     pass
